lumberjack 2 updates
