+++ CHANGE_LOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0
+
+### Changed
+
+- Bumped lumberjack version to >= 2.0.
+
+### Removed
+
+- Removed support for Ruby < 2.7.
+- Removed dependency on multi_json.
+
 ## 1.0.2
 
 ## Fixed
