+++ Gemfile
@@ -2,6 +2,8 @@ source "https://rubygems.org"
 
 gemspec
 
+gem "lumberjack", github: "bdurand/lumberjack", branch: "lumberjack_2"
+
 gem "fakeredis"
 
 gem "dotenv"
