+++ lumberjack_redis_device.gemspec
@@ -32,10 +32,9 @@ Gem::Specification.new do |spec|
 
   spec.require_paths = ["lib"]
 
-  spec.required_ruby_version = ">=2.4"
+  spec.required_ruby_version = ">=2.7"
 
-  spec.add_dependency "lumberjack", ">=1.1"
-  spec.add_dependency "multi_json"
+  spec.add_dependency "lumberjack", ">=2.0"
   spec.add_dependency "redis"
 
   spec.add_development_dependency "bundler"
